refactor: change package name and add KDocs
Change package from `utils.representation` to `utils.placement_strategy`. Add KDocs for CircularPlacementStrategy and RandomPlacementStrategy classes. Recode RandomPlacementStrategy to more effective code.
